requirements.txt
Flask==3.0.0
Pillow==10.2.0
pycryptodome==3.20.0
